requirements.txt
Flask
Flask-CORS
google-generativeai